requirements.txt
python-telegram-bot==13.15
pymongo==4.3.3
python-dotenv==1.0.0
aiohttp==3.8.4
motor==3.1.1
pytz==2023.3
babel==2.12.1
captcha==0.4
pillow==9.5.0